reorganize utility code, add noHueristics flag
